Split subnet matcher
